+++ Source/AIAccountListPreferencesPlugin.m
@@ -39,6 +39,15 @@
 }
 
 /*!
+ * @brief Uninstall
+ */
+- (void)uninstallPlugin
+{
+	// Remove the observer installPlugin registered so an uninstalled plugin stops receiving notifications.
+	[[NSNotificationCenter defaultCenter] removeObserver:self name:@"AIEditAccount" object:nil];
+}
+
+/*!
  * @brief Edit an account
  *
  * @param inNotification An AIEditAccount notification whose object is the AIAccount to edit