+++ Source/AILoggerPlugin.h
@@ -59,6 +59,8 @@
 	BOOL isIndexing;
 	SInt64 logsToIndex;
 	SInt64 logsIndexed;
+
+	NSToolbarItem *registeredToolbarItem;
 }
 @property(assign, readonly) BOOL indexIsFlushing;
 @property(assign, readonly) BOOL isIndexing;