+++ Source/BGEmoticonMenuPlugin.m
@@ -96,6 +96,12 @@
 {
 	[[NSNotificationCenter defaultCenter] removeObserver:self];
 	[adium.preferenceController unregisterPreferenceObserver:self];
+
+	// Unregister the toolbar item installPlugin registered, so an uninstalled plugin leaves no dead item behind.
+	if (registeredToolbarItem != nil) {
+		[adium.toolbarController unregisterToolbarItem:registeredToolbarItem forToolbarType:@"TextEntry"];
+		registeredToolbarItem = nil;
+	}
 }
 
 /*!
@@ -157,6 +163,7 @@
 														   menu:nil];
 	[button setToolbarItem:toolbarItem];
 	[adium.toolbarController registerToolbarItem:toolbarItem forToolbarType:@"TextEntry"];
+	registeredToolbarItem = toolbarItem;
 }
 
 // Menu Generation