+++ Tests/CoverageHost/AdiumY/AIPreferencePane.h
@@ -1,0 +1,1 @@
+../../../Frameworks/Adium/Source/AIPreferencePane.h